+++ node/src/BaseClient.ts
@@ -673,7 +673,7 @@ function toProtobufRoute(
             if (split.length !== 2) {
                 throw new RequestError(
                     "No port provided, expected host to be formatted as `{hostname}:{port}`. Received " +
-                    host,
+                        host,
                 );
             }
 
@@ -915,7 +915,8 @@ export class BaseClient {
             | command_request.Command
             | command_request.Command[]
             | command_request.ScriptInvocation
-            | command_request.ClusterScan,
+            | command_request.ClusterScan
+            | command_request.UpdateConnectionPassword,
         options: WritePromiseOptions = {},
     ): Promise<T> {
         const route = toProtobufRoute(options?.route);
@@ -985,35 +986,36 @@ export class BaseClient {
             | command_request.Command
             | command_request.Command[]
             | command_request.ScriptInvocation
-            | command_request.ClusterScan,
+            | command_request.ClusterScan
+            | command_request.UpdateConnectionPassword,
         route?: command_request.Routes,
     ) {
         const message = Array.isArray(command)
             ? command_request.CommandRequest.create({
-                callbackIdx,
-                transaction: command_request.Transaction.create({
-                    commands: command,
-                }),
-            })
+                  callbackIdx,
+                  transaction: command_request.Transaction.create({
+                      commands: command,
+                  }),
+              })
             : command instanceof command_request.Command
-                ? command_request.CommandRequest.create({
+              ? command_request.CommandRequest.create({
                     callbackIdx,
                     singleCommand: command,
                 })
-                : command instanceof command_request.ClusterScan
-                    ? command_request.CommandRequest.create({
+              : command instanceof command_request.ClusterScan
+                ? command_request.CommandRequest.create({
+                      callbackIdx,
+                      clusterScan: command,
+                  })
+                : command instanceof command_request.UpdateConnectionPassword
+                  ? command_request.CommandRequest.create({
                         callbackIdx,
-                        clusterScan: command,
+                        updateConnectionPassword: command,
                     })
-                    : command instanceof command_request.UpdateConnectionPassword
-                        ? command_request.CommandRequest.create({
-                            callbackIdx,
-                            updateConnectionPassword: command,
-                        })
-                        : command_request.CommandRequest.create({
-                            callbackIdx,
-                            scriptInvocation: command,
-                        });
+                  : command_request.CommandRequest.create({
+                        callbackIdx,
+                        scriptInvocation: command,
+                    });
         message.route = route;
 
         this.writeOrBufferRequest(
@@ -5989,9 +5991,9 @@ export class BaseClient {
         ReadFrom,
         connection_request.ReadFrom
     > = {
-            primary: connection_request.ReadFrom.Primary,
-            preferReplica: connection_request.ReadFrom.PreferReplica,
-        };
+        primary: connection_request.ReadFrom.Primary,
+        preferReplica: connection_request.ReadFrom.PreferReplica,
+    };
 
     /**
      * Returns the number of messages that were successfully acknowledged by the consumer group member of a stream.
@@ -7300,8 +7302,8 @@ export class BaseClient {
             res === null
                 ? null
                 : res!.map((r) => {
-                    return { key: r.key, elements: r.value };
-                })[0],
+                      return { key: r.key, elements: r.value };
+                  })[0],
         );
     }
 
@@ -7343,8 +7345,8 @@ export class BaseClient {
             res === null
                 ? null
                 : res!.map((r) => {
-                    return { key: r.key, elements: r.value };
-                })[0],
+                      return { key: r.key, elements: r.value };
+                  })[0],
         );
     }
 
@@ -7551,11 +7553,11 @@ export class BaseClient {
             : connection_request.ReadFrom.Primary;
         const authenticationInfo =
             options.credentials !== undefined &&
-                "password" in options.credentials
+            "password" in options.credentials
                 ? {
-                    password: options.credentials.password,
-                    username: options.credentials.username,
-                }
+                      password: options.credentials.password,
+                      username: options.credentials.username,
+                  }
                 : undefined;
         const protocol = options.protocol as
             | connection_request.ProtocolVersion
@@ -7676,5 +7678,47 @@ export class BaseClient {
             socket.end();
             throw err;
         }
+    }
+
+    /**
+     * Update the current connection with a new password.
+     *
+     * This method is useful in scenarios where the server password has changed or when utilizing short-lived passwords for enhanced security.
+     * It allows the client to update its password to reconnect upon disconnection without the need to recreate the client instance.
+     * This ensures that the internal reconnection mechanism can handle reconnection seamlessly, preventing the loss of in-flight commands.
+     *
+     * This method updates the client's internal password configuration and does not perform password rotation on the server side.
+     *
+     * @param password - The new password to update the current password, or `null` to remove the current password.
+     * @param reAuth - If `true`, the client will re-authenticate immediately with the new password. If `false`, the new password will be used for the next connection attempt.
+     * @returns Always `"OK"`.
+     *
+     * @example
+     * ```typescript
+     * await client.updateConnectionPassword("newPassword", true) // "OK"
+     * ```
+     */
+    async updateConnectionPassword(password: string | null, reAuth: boolean) {
+        const updateConnectionPassword =
+            command_request.UpdateConnectionPassword.create({
+                password: password,
+                reAuth,
+            });
+
+        const response = await this.createWritePromise<GlideString>(
+            updateConnectionPassword,
+        );
+
+        if (response === "OK" && !this.config?.credentials) {
+            this.config = {
+                ...this.config!,
+                credentials: {
+                    ...this.config!.credentials,
+                    password: password ? password : "",
+                },
+            };
+        }
+
+        return response;
     }
 }